pyproject.toml
[build-system]
requires = [
    "flit_core >=3.2,<4",
]
build-backend = "flit_core.buildapi"


[project]
name = "sheepy"
version = "0.0.1"
authors = [{name=  "Jannes Hachmer", email= "jannes.hachmer@gmail.com"}]
description = "add movies to your google spreadsheet"
readme = "README.md"
requires-python = ">3.11"
dependencies = [
    "gspread",
    "tabulate",
    "python-dotenv"
]


[tool.black]
line-length = 88
